Move type conversion functions to io module
