+++ tests/test_mine_equivalence_classes.py
@@ -2,7 +2,7 @@ import pytest
 import numpy as np
 
 from caspailleur import mine_equivalence_classes as mec
-from caspailleur import base_functions as bfunc
+from caspailleur import io
 
 from bitarray import frozenbitarray as fbarray
 from bitarray.util import zeros as bazeros
@@ -16,14 +16,14 @@ def test_list_intents_via_LCM():
         {1, 2, 3}
     ]
 
-    intents_true = list(bfunc.isets2bas([set(), {0}, {2}, {3}, {0, 2}, {0, 3}, {1, 2}, {1, 2, 3}, {0, 1, 2, 3, 4}], 5))
+    intents_true = list(io.isets2bas([set(), {0}, {2}, {3}, {0, 2}, {0, 3}, {1, 2}, {1, 2, 3}, {0, 1, 2, 3, 4}], 5))
 
-    intents = mec.list_intents_via_LCM(list(bfunc.isets2bas(itemsets, 5)))
+    intents = mec.list_intents_via_LCM(list(io.isets2bas(itemsets, 5)))
     assert intents == intents_true
 
     K = np.array([[True, False, False, False, False, True], [False, True, False, False, False, False]])
-    itemsets = bfunc.np2bas(K)
-    intents_true = list(bfunc.isets2bas([set(), {1}, {0, 5}, {0, 1, 2, 3, 4, 5}], 6))
+    itemsets = io.np2bas(K)
+    intents_true = list(io.isets2bas([set(), {1}, {0, 5}, {0, 1, 2, 3, 4, 5}], 6))
 
     intents = mec.list_intents_via_LCM(itemsets)
     assert intents == intents_true
@@ -36,8 +36,8 @@ def test_list_intents_via_lindig():
     [False, True, True, False],
     [False, True, True, True]])
 
-    itemsets = bfunc.np2bas(K)
-    attr_extents = bfunc.np2bas(K.T)
+    itemsets = io.np2bas(K)
+    attr_extents = io.np2bas(K.T)
     intents_true = ['1111', '1001', '1010', '0111', '1000', '0010', '0001', '0000', '0110']
  
     intents_true = [bitarray(x) for x in intents_true]
@@ -50,7 +50,7 @@ def test_list_attribute_concepts():
     intents = [set(), {0}, {2}, {3}, {0, 2}, {0, 3}, {1, 2}, {1, 2, 3},  {0, 1, 2, 3, 4}]
     attr_concepts_true = [1, 6, 2, 3, 8]
 
-    attr_concepts = mec.list_attribute_concepts(list(bfunc.isets2bas(intents, 5)))
+    attr_concepts = mec.list_attribute_concepts(list(io.isets2bas(intents, 5)))
     assert attr_concepts == attr_concepts_true
 
 
@@ -61,7 +61,7 @@ def tests_iter_equivalence_class():
         [False, True, True, False, False],
         [False, True, True, True, False],
     ])
-    attr_extents = list(bfunc.np2bas(K.T))
+    attr_extents = list(io.np2bas(K.T))
     eq_class_true = [
         {0, 1, 2, 3, 4}, {0, 1, 2, 3}, {0, 1, 2, 4}, {0, 1, 3, 4}, {0, 2, 3, 4}, {1, 2, 3, 4},
         {0, 1, 2}, {0, 1, 3}, {0, 1, 4}, {0, 2, 3}, {0, 2, 4}, {0, 3, 4}, {1, 2, 4}, {1, 3, 4}, {2, 3, 4},
@@ -69,16 +69,16 @@ def tests_iter_equivalence_class():
     ]
 
     eq_class = list(mec.iter_equivalence_class(attr_extents, ~fbarray(bazeros(5))))
-    assert eq_class == list(bfunc.isets2bas(eq_class_true, 5))
+    assert eq_class == list(io.isets2bas(eq_class_true, 5))
 
 
 def test_list_keys_via_eqclass():
-    eq_class = list(bfunc.isets2bas([
+    eq_class = list(io.isets2bas([
         {0, 1, 2, 3, 4}, {0, 1, 2, 3}, {0, 1, 2, 4}, {0, 1, 3, 4}, {0, 2, 3, 4}, {1, 2, 3, 4},
         {0, 1, 2}, {0, 1, 3}, {0, 1, 4}, {0, 2, 3}, {0, 2, 4}, {0, 3, 4}, {1, 2, 4}, {1, 3, 4}, {2, 3, 4},
         {0, 1}, {0, 4}, {1, 4}, {2, 4}, {3, 4}, {4}
     ], 5))
-    keys_true = list(bfunc.isets2bas([{0, 2, 3}, {0, 1}, {4}], 5))
+    keys_true = list(io.isets2bas([{0, 2, 3}, {0, 1}, {4}], 5))
 
     keys = mec.list_keys_via_eqclass(eq_class)
     assert keys == keys_true
@@ -88,14 +88,14 @@ def test_list_passkeys_via_keys():
     keys = [{0, 2, 3}, {0, 1}, {4}, {0, 2, 4}]
     pkeys_true = [{4}]
 
-    pkeys = mec.list_passkeys_via_eqclass(bfunc.isets2bas(keys, 5))
-    assert pkeys == list(bfunc.isets2bas(pkeys_true, 5))
+    pkeys = mec.list_passkeys_via_eqclass(io.isets2bas(keys, 5))
+    assert pkeys == list(io.isets2bas(pkeys_true, 5))
 
 
 def test_list_keys():
     n_attrs = 5
     intents = [set(), {0}, {2}, {3}, {0, 2}, {0, 3}, {1, 2}, {1, 2, 3}, {0, 1, 2, 3, 4}]
-    intents = list(bfunc.isets2bas(intents, n_attrs))
+    intents = list(io.isets2bas(intents, n_attrs))
 
     keys_true = [
         (set(), 0),
@@ -103,7 +103,7 @@ def test_list_keys():
         ({0, 1}, 8), ({0, 2}, 4), ({0, 3}, 5), ({1, 3}, 7), ({2, 3}, 7),
         ({0, 2, 3}, 8)
     ]
-    keys_true = {next(bfunc.isets2bas([key], n_attrs)): intent_i for key, intent_i in keys_true}
+    keys_true = {next(io.isets2bas([key], n_attrs)): intent_i for key, intent_i in keys_true}
 
     keys = mec.list_keys(intents)
     assert keys == keys_true
@@ -115,14 +115,14 @@ def test_list_keys():
 def test_list_passkeys():
     n_attrs = 5
     intents = [set(), {0}, {2}, {3}, {0, 2}, {0, 3}, {1, 2}, {1, 2, 3}, {0, 1, 2, 3, 4}]
-    intents = list(bfunc.isets2bas(intents, n_attrs))
+    intents = list(io.isets2bas(intents, n_attrs))
 
     pkeys_true = [
         (set(), 0),
         ({0}, 1), ({1}, 6), ({2}, 2), ({3}, 3), ({4}, 8),
         ({0, 2}, 4), ({0, 3}, 5), ({1, 3}, 7), ({2, 3}, 7),
     ]
-    pkeys_true = {next(bfunc.isets2bas([key], n_attrs)): intent_i for key, intent_i in pkeys_true}
+    pkeys_true = {next(io.isets2bas([key], n_attrs)): intent_i for key, intent_i in pkeys_true}
 
     pkeys = mec.list_passkeys(intents)
     assert pkeys == pkeys_true
